move test code into integration module
