+++ .github/workflows/test.yml
@@ -15,9 +15,7 @@ jobs:
       - uses: actions-rs/toolchain@v1
       - uses: Swatinem/rust-cache@v1
       - uses: actions-rs/cargo@v1
-        with:
-          command: test
-          args: --lib
+      - run: make test
 
   sdk-test:
     runs-on: ubuntu-latest
